Отделено движение шара в отдельный скрипт #72
