+++ Assets/Prefabs/LevelKit.prefab
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:61f528c3004ba994c0edfbc64600dafad84bd81f3e20e06f52d85960f1e6c549
-size 104160
+oid sha256:45aff301174a1fa8772fdc3cc324295ba9e4b63974e4068be8aeb62501f92502
+size 104600
