+++ Assets/Scenes/4 Beds and Fire.unity
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:ff1389af9ab7e1fe3073672cb27bc7b704543d130ff83acc4beeaae2628c6871
-size 19690
+oid sha256:bb74bab0fbe3f659969fe5efe0601e2338a51329975d7c265df2d16146216bc4
+size 18778
